backend/requirements.txt
# ==========================================
# 1. Web 后端基础框架
# ==========================================
Flask==3.0.2
Flask-Cors==4.0.0
Werkzeug==3.0.1

# ==========================================
# 2. API 通信与配置管理 (对接腾讯开悟、得理大模型必备)
# ==========================================
requests==2.31.0
python-dotenv==1.0.1  # 用于管理 API Key 等敏感配置，避免泄露到 Git

# ==========================================
# 3. 法律文书排版与生成
# ==========================================
docxtpl==0.16.7
google-generativeai>=0.4.0
python-dotenv==1.0.1

# ==========================================
# 4. 视觉识别与 OCR (YOLO 与飞桨)
# ==========================================
# 基础图像处理
opencv-python>=4.9.0
Pillow>=10.2.0

# YOLO 官方库 (放宽版本限制，适配新版 Python)
ultralytics>=8.2.0

# PaddleOCR 及其底层依赖 (根据你的报错提示，最低可用版本为 2.6.2)
paddlepaddle>=2.6.2
paddleocr>=2.7.3